Rename message-to-wallet as wallet-instruction
